misc-upgrade:     1. change --force --post to -f -p     2. add -l support for mass storage     3. change prepare and upgrade image     4. update readme

+++ package/softwinner/misc-upgrade/aw_upgrade_process.sh
@@ -13,6 +13,10 @@ USR_IMG=usr_sys.tar.gz
 
 UPGRADE_SH=/sbin/aw_upgrade_image.sh
 
+IMG_PATH=/mnt/UDISK/misc-upgrade/
+
+UPGRADE_SETTING_PATH=/mnt/UDISK/misc-upgrade/
+
 check_ip_timeout_default(){
     #$1 timeout(s)
     let timeout=0
@@ -95,7 +99,7 @@ download_prepare_image(){
     download_image $1 /tmp
     #md5 check for the package
     #if check ok
-    $UPGRADE_SH prepare /tmp/$1
+    $UPGRADE_SH prepare /tmp $1
 }
 boot_recovery_mode(){
     # boot-reocvery mode
@@ -105,29 +109,42 @@ boot_recovery_mode(){
 
     # umount the usr partition; if failed, ignore
     umount /usr
-   
+    umount /usr
+
     #try to mount rootfs_data partition
     try_mount "rootfs_data" "/overlay"
     try_mount "UDISK" "/mnt/UDISK"
 
-    # get current wifi wpa_supplicant.conf
-    [ -f /overlay/etc/wifi/wpa_supplicant.conf ] && {
-        echo "get wpa_supplicant from overlay"
-        echo "old wpa_supplicant.config"
-        cat /etc/wifi/wpa_supplicant.conf
-        cp /overlay/etc/wifi/wpa_supplicant.conf /etc/wifi/
-        echo "new wpa_supplicant.config"
-        cat /etc/wifi/wpa_supplicant.conf
-    }
-
-    check_network
-
     $UPGRADE_SH clean
-    
-    download_prepare_image $TARGET_IMG
 
     upgrade_start boot_recovery
-    $UPGRADE_SH upgrade
+    
+    IMG_PATH=`cat $UPGRADE_SETTING_PATH/.image_path`
+    if [ -f $IMG_PATH/$TARGET_IMG ]; then
+        $UPGRADE_SH prepare $IMG_PATH $TARGET_IMG
+        $UPGRADE_SH prepare $IMG_PATH $USR_IMG
+        $UPGRADE_SH upgrade
+    else
+        # get current wifi wpa_supplicant.conf
+        [ -f /overlay/etc/wifi/wpa_supplicant.conf ] && {
+            echo "get wpa_supplicant from overlay"
+            echo "old wpa_supplicant.config"
+            cat /etc/wifi/wpa_supplicant.conf
+            cp /overlay/etc/wifi/wpa_supplicant.conf /etc/wifi/
+            echo "new wpa_supplicant.config"
+            cat /etc/wifi/wpa_supplicant.conf
+        }
+
+        check_network
+
+        download_prepare_image $TARGET_IMG
+        $UPGRADE_SH upgrade
+        download_prepare_image $USR_IMG
+        $UPGRADE_SH upgrade
+    fi
+
+    #upgrade_start boot_recovery
+    #$UPGRADE_SH upgrade
     upgrade_finish
 }
 upgrade_pre_mode(){
@@ -139,19 +156,33 @@ upgrade_pre_mode(){
 
     # umount the usr partition; if failed, ignore
     umount /usr
-    
-    check_network
+    umount /usr
 
     $UPGRADE_SH clean
-    
-    download_prepare_image $RAMDISK_IMG
-    download_prepare_image $TARGET_IMG
-    
-    upgrade_start pre
 
-    if [ $? -eq 0 ] || [ ! x$NORMAL_MODE = x"normal" ]; then
+    upgrade_start pre
+    
+    IMG_PATH=`cat $UPGRADE_SETTING_PATH/.image_path`
+    if [ -f $IMG_PATH/$RAMDISK_IMG ] && [ -f $IMG_PATH/$TARGET_IMG ]; then
+        $UPGRADE_SH prepare $IMG_PATH $RAMDISK_IMG
+        $UPGRADE_SH prepare $IMG_PATH $TARGET_IMG
+        $UPGRADE_SH prepare $IMG_PATH $USR_IMG
+        $UPGRADE_SH upgrade
+    else
+        check_network
+        download_prepare_image $RAMDISK_IMG
+        $UPGRADE_SH upgrade
+        download_prepare_image $TARGET_IMG
+        $UPGRADE_SH upgrade
+        download_prepare_image $USR_IMG
         $UPGRADE_SH upgrade
     fi
+
+    #upgrade_start pre
+
+    #if [ $? -eq 0 ] || [ ! x$NORMAL_MODE = x"normal" ]; then
+    #    $UPGRADE_SH upgrade
+    #fi
     upgrade_finish
 }
 upgrade_post_mode(){
@@ -160,12 +191,19 @@ upgrade_post_mode(){
     # ---->write image to "extend" partition 
     echo "upgrade_post_mode"
 
-    check_network
+    umount /usr
+    umount /usr
 
     $UPGRADE_SH clean
     
-    download_prepare_image $USR_IMG
- 
+    IMG_PATH=`cat $UPGRADE_SETTING_PATH/.image_path`
+    if [ -f $IMG_PATH/$USR_IMG ]; then
+        $UPGRADE_SH prepare $IMG_PATH $USR_IMG
+    else
+        check_network
+        download_prepare_image $USR_IMG
+    fi
+
     upgrade_start post
     if [ $? -eq 0 ] || [ ! x$NORMAL_MODE = x"normal" ]; then
         $UPGRADE_SH upgrade
@@ -176,16 +214,47 @@ upgrade_end_mode(){
     # upgrade_end mode
     # wait for next upgrade
     echo "wait for next upgrade!!"
+    #clear
+    [ -f $UPGRADE_SETTING_PATH/.image_path ] && rm -rf $UPGRADE_SETTING_PATH/.image_path
 }
 ####################################################
+modeflag=0
+check_mode(){
+    [ $modeflag -eq 1 ] && {
+        echo "mode conflict, must be set -p or -f"
+        exit 0
+    }
+    modeflag=1
+}
+while getopts "fpl:" opt; do
+    case $opt in
+    f)
+        check_mode
+        mode="--force"
+        ;;
+    p)
+        check_mode
+        mode="--post"
+        ;;
+    l)
+        [ ! -d $OPTARG ] && echo "-l $OPTARG, the settting path is unavailable" && exit 0
+        mkdir -p $UPGRADE_SETTING_PATH
+        echo $OPTARG > $UPGRADE_SETTING_PATH/.image_path
+        ;;  
+    \?)  
+        echo "Invalid option: -$OPTARG"
+        exit 0
+        ;;  
+    esac  
+done  
 # force to upgrade
-if [ -n $1 ] && [ x$1 = x"--force" ]; then
+if [ -n $mode ] && [ x$mode = x"--force" ]; then
     export NORMAL_MODE=normal
     upgrade_pre_mode
     exit 0
 fi
 
-if [ -n $1 ] && [ x$1 = x"--post" ]; then
+if [ -n $mode ] && [ x$mode = x"--post" ]; then
     export NORMAL_MODE=normal
     upgrade_post_mode
     exit 0
